requirements.txt
gpxpy
matplotlib
moviepy
numpy
pandas
pillow
proglog
py-staticmaps[cairo]
